# performance: fuse more vectorized operations